+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 - #1370 - Fixed Invalid Entries display in FireFox
 - #1371 - Fixed Incorrect Entry Deletion when Filtering
 
+### Added
+- #1365 - Sling model injector for Shared Component Property values.
+
 [Unreleased]: https://github.com/Adobe-Consulting-Services/acs-aem-commons/compare/acs-aem-commons-3.16.0...HEAD
 
 ## [3.16.0] - 2018-05-10